[EGD-7947] Unified aproach for download_asset
Improved timing. Uses cached file for both json and direct
downloads if possible.
Improved logging. Clear information what's being downloaded.
Removed older script.
